ai-microservice/requirements.txt
# FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# OpenAI and AI models
openai>=1.50.0
anthropic>=0.7.0
langchain>=0.0.340
langchain-openai>=0.0.2
langchain-anthropic>=0.1.0

# Music notation and processing
music21>=9.1.0

# Database and caching
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
redis>=5.0.1

# Image processing
Pillow>=10.1.0

# HTTP requests and web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Audio processing
numpy<2.0.0
scipy>=1.11.0
librosa>=0.11.0
soundfile>=0.13.0
basic-pitch==0.4.0
crepe==0.0.16
tensorflow-cpu==2.14.0

# PDF processing (for IMSLP)
PyPDF2>=3.0.0

# Firebase (for IMSLP image storage)
firebase-admin>=6.0.0

# Data processing
pydantic>=2.5.0
python-dotenv>=1.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
